Pull migration code into its own package
