+++ collect_app/src/main/java/org/odk/collect/android/activities/SplashScreenActivity.java
@@ -29,7 +29,7 @@ import android.widget.LinearLayout;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.analytics.Analytics;
-import org.odk.collect.android.application.ApplicationInitializer;
+import org.odk.collect.android.application.initialization.ApplicationInitializer;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.listeners.PermissionListener;
 import org.odk.collect.android.preferences.GeneralKeys;
