+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/migration/KeyRenamer.java
@@ -1,0 +1,27 @@
+package org.odk.collect.android.application.initialization.migration;
+
+import android.content.SharedPreferences;
+
+import static org.odk.collect.android.application.initialization.migration.MigrationUtils.replace;
+
+public class KeyRenamer implements Migration {
+
+    String oldKey;
+    String newKey;
+
+    KeyRenamer(String oldKey) {
+        this.oldKey = oldKey;
+    }
+
+    public KeyRenamer toKey(String newKey) {
+        this.newKey = newKey;
+        return this;
+    }
+
+    public void apply(SharedPreferences prefs) {
+        if (prefs.contains(oldKey) && !prefs.contains(newKey)) {
+            Object value = prefs.getAll().get(oldKey);
+            replace(prefs, oldKey, newKey, value);
+        }
+    }
+}
