+++ collect_app/src/main/java/org/odk/collect/android/application/initialization/migration/Migration.java
@@ -1,0 +1,7 @@
+package org.odk.collect.android.application.initialization.migration;
+
+import android.content.SharedPreferences;
+
+public interface Migration {
+    void apply(SharedPreferences prefs);
+}
